refactor: extract gh.lua, highlights.lua, and split large modules
- Add lua/gh_dashboard/gh.lua: single shared async gh CLI runner, replacing
  4 duplicate run_gh() implementations across modules
- Add lua/gh_dashboard/highlights.lua: centralise all 57 highlight group
  definitions with a single ColorScheme autocmd and registered guard
- Split lua/gh_dashboard/init.lua (was ~946 lines) into:
  - dashboard/fetch.lua: all 9 API fetch functions
  - dashboard/render.lua: all panel render functions + build() entry point
  - init.lua: thin orchestrator (~299 lines)
- Split lua/gh_dashboard/reader.lua (was ~1082 lines) into:
  - reader/fetch.lua: fetch_issue, fetch_pr, fetch_diff, fetch_readme, etc.
  - reader/render.lua: markdown body, diff, CI, review renderers (pure)
  - reader/actions.lua: post_comment, submit_review, merge_pr, close_issue
  - reader/init.lua: state, window management, keymaps, public M.open/open_diff
- Update watchlist.lua, user_profile.lua, user_watchlist.lua to use shared modules
