+++ .specify/templates/constitution-template.md
@@ -1,0 +1,62 @@
+# Project Constitution: [PROJECT_NAME]
+
+**Version**: [CONSTITUTION_VERSION]
+**Ratified**: [RATIFICATION_DATE]
+**Last Amended**: [LAST_AMENDED_DATE]
+**Maintainer**: [MAINTAINER]
+
+---
+
+## Purpose
+
+[PURPOSE — one paragraph describing what this project does and who it serves]
+
+---
+
+## Principles
+
+### [PRINCIPLE_1_NAME]
+
+[PRINCIPLE_1_BODY — describe the non-negotiable rules and rationale]
+
+### [PRINCIPLE_2_NAME]
+
+[PRINCIPLE_2_BODY]
+
+### [PRINCIPLE_3_NAME]
+
+[PRINCIPLE_3_BODY]
+
+### [PRINCIPLE_4_NAME]
+
+[PRINCIPLE_4_BODY]
+
+---
+
+## Scope & Boundaries
+
+**In scope:**
+- [SCOPE_IN_1]
+- [SCOPE_IN_2]
+
+**Out of scope:**
+- [SCOPE_OUT_1]
+- [SCOPE_OUT_2]
+
+---
+
+## Governance
+
+### Amendment Procedure
+
+[AMENDMENT_PROCEDURE — how changes to this document are proposed, reviewed, and ratified]
+
+### Versioning Policy
+
+- MAJOR: Backward-incompatible governance or principle removals/redefinitions
+- MINOR: New principle or section added or materially expanded
+- PATCH: Clarifications, wording, typo fixes
+
+### Compliance Review
+
+[COMPLIANCE_REVIEW — when and how compliance with these principles is checked]
