+++ .specify/templates/plan-template.md
@@ -1,0 +1,78 @@
+# Implementation Plan: [FEATURE_NAME]
+
+**Branch**: [BRANCH_NAME]
+**Spec**: [SPEC_FILE]
+**Created**: [DATE]
+
+---
+
+## Constitution Check
+
+> Verify this plan complies with project principles before proceeding.
+
+| Principle | Compliant? | Notes |
+|-----------|-----------|-------|
+| [PRINCIPLE_1] | ✅ / ⚠️ / ❌ | [NOTES] |
+| [PRINCIPLE_2] | ✅ / ⚠️ / ❌ | [NOTES] |
+
+---
+
+## Technical Context
+
+**Affected modules**: [LIST]
+**New modules**: [LIST or none]
+**External dependencies**: [LIST or none]
+**Breaking changes**: [YES/NO — describe if yes]
+
+---
+
+## Research Summary
+
+> See research.md for full decisions. Key choices:
+
+- [DECISION_1]: [RATIONALE]
+- [DECISION_2]: [RATIONALE]
+
+---
+
+## Data Model Changes
+
+> See data-model.md if applicable.
+
+[SUMMARY or "No data model changes"]
+
+---
+
+## Implementation Phases
+
+### Phase 1: [PHASE_1_NAME]
+
+**Goal**: [WHAT THIS PHASE ACHIEVES]
+**Deliverable**: [TESTABLE OUTCOME]
+
+Tasks:
+- [ ] [TASK_1]
+- [ ] [TASK_2]
+
+### Phase 2: [PHASE_2_NAME]
+
+**Goal**: [WHAT THIS PHASE ACHIEVES]
+**Deliverable**: [TESTABLE OUTCOME]
+
+Tasks:
+- [ ] [TASK_3]
+- [ ] [TASK_4]
+
+---
+
+## Rollout & Verification
+
+[HOW TO VERIFY THE FEATURE WORKS END-TO-END]
+
+---
+
+## Risks & Mitigations
+
+| Risk | Likelihood | Impact | Mitigation |
+|------|-----------|--------|-----------|
+| [RISK_1] | Low/Med/High | Low/Med/High | [MITIGATION] |
