+++ .specify/templates/spec-template.md
@@ -1,0 +1,89 @@
+# Feature Specification: [FEATURE_NAME]
+
+**Status**: Draft
+**Created**: [DATE]
+**Branch**: [BRANCH_NAME]
+
+---
+
+## Overview
+
+[One paragraph: what this feature does and why it matters to users/contributors]
+
+---
+
+## Problem Statement
+
+[Describe the current pain point or gap this feature addresses. Written from the user's perspective, not the developer's.]
+
+---
+
+## Goals
+
+- [GOAL_1]
+- [GOAL_2]
+- [GOAL_3]
+
+## Non-Goals
+
+- [NON_GOAL_1]
+- [NON_GOAL_2]
+
+---
+
+## User Scenarios & Testing
+
+### Primary Flow
+
+**Actor**: [WHO]
+**Trigger**: [WHAT INITIATES THIS]
+**Steps**:
+1. [STEP_1]
+2. [STEP_2]
+3. [STEP_3]
+**Outcome**: [WHAT THE USER ACHIEVES]
+
+### Edge Cases
+
+- [EDGE_CASE_1]
+- [EDGE_CASE_2]
+
+---
+
+## Functional Requirements
+
+### Must Have (P1)
+
+- [ ] [REQUIREMENT_1]
+- [ ] [REQUIREMENT_2]
+
+### Should Have (P2)
+
+- [ ] [REQUIREMENT_3]
+
+### Nice to Have (P3)
+
+- [ ] [REQUIREMENT_4]
+
+---
+
+## Success Criteria
+
+| Criterion | Measure | Target |
+|-----------|---------|--------|
+| [CRITERION_1] | [HOW_MEASURED] | [TARGET_VALUE] |
+| [CRITERION_2] | [HOW_MEASURED] | [TARGET_VALUE] |
+
+---
+
+## Assumptions
+
+- [ASSUMPTION_1]
+- [ASSUMPTION_2]
+
+---
+
+## Dependencies
+
+- [DEPENDENCY_1]
+- [DEPENDENCY_2]
